+++ philo.h
@@ -40,7 +40,7 @@ typedef struct s_rules
 	int				finish;
 	pthread_mutex_t	*forks;
 	pthread_mutex_t	write;
-	uint16_t		start_time;
+	uint64_t		start_time;
 	pthread_t		monitoring_thread;
 	t_philo			*philo;
 }				t_rules;
@@ -50,6 +50,6 @@ void		ft_putstr(char *str);
 int			ft_isnumeric(int argc, char *argv[]);
 int			ft_atoi(const char *str);
 void		ft_threadmaker(t_rules *rules);
-void		print_msg(uint16_t start_time, int id, pthread_mutex_t	*write, char *str);
-uint16_t	get_time(void);
+void		print_msg(uint64_t start_time, int id, pthread_mutex_t	*write, char *str);
+uint64_t	get_time(void);
 #endif
